pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = "Template for FastAPI projects"
authors = [ "Leonardo Alvarez <leonardo.alvarez.dev.20@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.1"
SQLAlchemy = "^1.4.42"
pendulum = "^2.1.2"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.5"
retrying = "^1.3.3"
secure = "^0.3.0"
uvicorn = "^0.18.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
ipython = "^8.5.0"
pytest-cov = "^4.0.0"
pytest-sugar = "^0.9.5"
mock = "^4.0.3"
factory-boy = "^3.2.1"
pre-commit = "^2.20.0"
commitizen = "^2.37.0"
requests = "^2.28.1"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 80

[tool.black]
line-length = 80
exclude = '''
  /(
      \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  '''

[tool.interrogate]
ignore-init-module = true
ignore-init-method = true
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-module = false
fail-under = 50
exclude = ["app/utils", "app/config"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = true
quiet = false
whitelist-regex = []
color = true

[tool.vulture]
ignore_decorators = ["@app.route", "@require_*"]
paths = ["app", "tests"]
ignore_names = ["dialect"]
min_confidence = 80
sort_by_size = true
verbose = false

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.3.3"
tag_format = "v$major.$minor.$patch"
version_files = [
    "app/__init__.py",
    "pyproject.toml:version"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
